.bashrc rem: With no args, list push URLs along with the remotes
It's nice to see the sources of the remotes as well as their names in
the default listing. But displaying both pull and push URLs, `git
remote -v` does, is noisy and most of the time we can get all the
information we need from the just the push URL. (Typically, if they're
different, the pull URL is from the same repo but just, e.g., without
authentication.) So given no arguments display the push URLs as well
(taking up no extra lines on the display) and now we don't have to
grovel through the twice as much output as `git remote -v` gives us.