.bashrc: Add __git_tools_gitver_GE and fix logab on Git < 1.8
Git < 1.8 (or maybe 1.9) doesn't have the `--exclude` option to `log`
and so this caused `logab` to fail when it tried to exclude notes.

Add a new `__git_tools_gitver_GE` function that returns false if the
installed Git is not at least the version given as a parameter. Use
this to drop the `--exclude` argument to `log` if Git < 1.8.