.bashrc pushu: Unset upstream before push
This avoids it using the previous upstream after a branch has been
renamed, which is usually not what we want. Example problem:

    $ git branch one
    $ pushu origin          # Create `one` on server
    $ br -m one two
    $ push origin :one
    $ git status -b -s
    ## two...origin/one     # Still tracking old name!
    $ pushu origin          # Re-creates `one` on server

For the moment, we hack this by always unsetting the upstream when
this command is used (which probably fails on older versions of Git).

We need to make this command work better, probably starting by
figuring out what DWIM would really mean for this command.