.bashrc: Import many more git functions from private repo
This is pretty much all the remaining stuff I use regularly.
There's a lot of cleanup etc. that needs to be done here,
particularly adding completion to more of the functions.

The following used to be aliases to my `inf` script which made them
work for CVS and Subversion as well; this isn't something I use any
more (I'm more likely to use `git-cvs` or `git-subversion to interact
with those repos) and so they've been replaced with simple shell
functions that work only for Git.

    log, dif, blame, add