.bashrc logbr(): Make argument parsing better, though not perfect
It now works and does the expected thing with no arguments.

It also checks the last argument to see if it starts with `-` and
assumes the branch is `@` if it does. This works if the last argument
is `--pretty=oneline` but not with `--pretty oneline`. I'm not sure
how to fix that.