.bashrc logbr(): Log with --stat information
I usually want this. It would be nice to have a way to turn it off,
but what we really want is a way to combine all the various `logbr`,
`logs`, `logm` etc. into one function where I could easily and
concisely specify the different bits of information I want to turn on
and off.