.bashrc {i}clean(): git clean with default `-n` and clean ignores
`clean` supplies the required `-n` (to display what will be cleaned)
if you don't explicitly supply a `-f` (to actually do the clean) in
the arguments.

`iclean` supplies `-dX` to clean ignored files and directories. This
doesn't touch untracked files and dirs that are not explicitly
ignored in a `.gitignore` or similar.