.bashrc mbase(): Work when primary remote is not 'origin'
The default second branch ref (from which we look for where we
branched off) was `origin/master`, which doesn't work if there's no
`origin` remote. We change that to `master@{upstream}` to always use
the branch our local master is tracking, regardless of the name of the
remote for that tracking branch.