.bashrc: Replace `grm` with `grmu` to rebase on remote 'master'
This has always rebased on a remote tracking branch rather than the
local `master` because we want to be able to fetch and rebase when
we're on a dev branch without having to update the local `master`
branch.

Previously we always rebased on `origin/master`, but that's not always
the proper tracking branch, especially if there's no remote named
`origin`. Now we figure out which remote's master to use by using
whatever the local `master` branch is tracking.