.bashrc: Make copy_git_completion work on Git for Windows, too
They keep the Bash script that sets up Git completion in a different
location, so to make sure we get it all we need to do is try sourcing
that one too.

Also we have to change the test that it worked from looking for the
existence of the particular script to confirming that the function we
wanted to load did get loaded, which is a better test anyway. That
lets this work even if the completion setup script is in neither
location so long as it was loaded before this code runs. (That would
presumably be done by the end user in the earlier setup for his
particular environment.)